+++ pilot/pkg/networking/core/v1alpha3/envoyfilter/listener_patch.go
@@ -29,6 +29,7 @@ import (
 	networking "istio.io/api/networking/v1alpha3"
 	"istio.io/istio/pilot/pkg/model"
 	"istio.io/istio/pilot/pkg/networking/util"
+	"istio.io/istio/pilot/pkg/util/goutils"
 )
 
 // ApplyListenerPatches applies patches to LDS output
@@ -38,7 +39,7 @@ func ApplyListenerPatches(
 	push *model.PushContext,
 	listeners []*xdsapi.Listener,
 	skipAdds bool) (out []*xdsapi.Listener) {
-	defer util.HandleCrash(func() {
+	defer goutils.HandleCrash(func() {
 		log.Errorf("listeners patch caused panic, so the patches did not take effect")
 	})
 	// In case the patches cause panic, use the listeners generated before to reduce the influence.
